WorldObjects now Tick properly
